Added http variable
